+++ setup.py
@@ -44,5 +44,5 @@ setuptools.setup(
         "Programming Language :: Python :: Implementation :: CPython",
         "Programming Language :: Python :: Implementation :: PyPy",
     ],
-    entry_points={"pytest11": ["twisted = pytest_twisted.core"]},
+    entry_points={"pytest11": ["twisted = pytest_twisted"]},
 )
